update(extract): construct the VO
next: trash judging
