+++ extract/judge_if_trash.py
@@ -1,0 +1,4 @@
+
+
+def predict(content):
+    return True
